net: Explode protocol::io module
Possibly good to split out into its own crate soon, to avoid unwieldy
visibility management.

Also note that the receive loop now propagates panics, but that's only a
small bandaid (see #505).
