+++ librad/src/net/protocol/io/recv.rs
@@ -3,18 +3,8 @@
 // This file is part of radicle-link, distributed under the GPLv3 with Radicle
 // Linking Exception. For full terms see the included LICENSE file.
 
-pub mod delegation;
-pub mod generic;
-pub mod git;
-pub mod payload;
-pub mod relations;
-pub mod sign;
-pub mod urn;
-pub use urn::Urn;
+mod gossip;
+pub(in crate::net::protocol) use gossip::gossip;
 
-mod sealed;
-
-#[cfg(test)]
-pub(crate) mod gen;
-
-pub use git::*;
+mod membership;
+pub(in crate::net::protocol) use membership::membership;
